requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests for FMCSA API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0

# Dashboard
streamlit==1.28.0
pandas==2.1.3
plotly==5.18.0

# Date handling
python-dateutil==2.8.2

# Testing (optional but recommended)
pytest==7.4.3
httpx==0.25.2
